renamed graphicContentModal.js to ConsentForm.js
